requirements.txt
python-telegram-bot>=21.0
anthropic>=0.39.0
notion-client>=2.2.1
requests>=2.32.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
